feat: add jwt auth (identity) beside social sign in - add identity based authentication. - remove redundant boilerplate code.
