+++ backend/DeltaShareApi/Data/AuthorizedToken.cs
@@ -1,0 +1,7 @@
+﻿namespace DeltaShareApi.Data;
+
+public class AuthorizedToken
+{
+    public Guid Id { get; set; }
+    public string Token { get; set; } = default!;
+}
