--- backend/DeltaShareApi/Data/AuthorizedTokens.cs
@@ -1,8 +1,0 @@
-﻿namespace DeltaShareApi.Data
-{
-    public class AuthorizedTokens
-    {
-        public Guid Id { get; set; }
-        public string Token { get; set; } = default!;
-    }
-}
