+++ backend/DeltaShareApi/Data/UserContext.cs
@@ -1,39 +1,39 @@
 ﻿using AvantiPoint.MobileAuth.Stores;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace DeltaShareApi.Data
+namespace DeltaShareApi.Data;
+
+public class UserContext : IdentityDbContext, ITokenStore
 {
-    public class UserContext : DbContext, ITokenStore
+    public UserContext(DbContextOptions options)
+        : base(options)
     {
-        public UserContext(DbContextOptions options)
-            : base(options)
+    }
+
+    public DbSet<AuthorizedToken> AuthorizedTokens { get; set; }
+
+    public DbSet<UserRole> UserRoles { get; set; }
+
+    public async ValueTask AddToken(string jwt, DateTimeOffset expires)
+    {
+        await AuthorizedTokens.AddAsync(new AuthorizedToken()
         {
-        }
+            Token = jwt
+        });
+        await SaveChangesAsync();
+    }
 
-        public DbSet<AuthorizedTokens> AuthorizedTokens { get; set; }
-
-        public DbSet<UserRole> UserRoles { get; set; }
-
-        public async ValueTask AddToken(string jwt, DateTimeOffset expires)
+    public async ValueTask RemoveToken(string jwt)
+    {
+        var tokens = await AuthorizedTokens.Where(x => x.Token == jwt).ToArrayAsync();
+        if (tokens.Any())
         {
-            await AuthorizedTokens.AddAsync(new AuthorizedTokens()
-            {
-                Token = jwt
-            });
+            AuthorizedTokens.RemoveRange(tokens);
             await SaveChangesAsync();
         }
-
-        public async ValueTask RemoveToken(string jwt)
-        {
-            var tokens = await AuthorizedTokens.Where(x => x.Token == jwt).ToArrayAsync();
-            if (tokens.Any())
-            {
-                AuthorizedTokens.RemoveRange(tokens);
-                await SaveChangesAsync();
-            }
-        }
-
-        public async ValueTask<bool> TokenExists(string jwt) =>
-            await AuthorizedTokens.AnyAsync(x => x.Token == jwt);
     }
+
+    public async ValueTask<bool> TokenExists(string jwt) =>
+        await AuthorizedTokens.AnyAsync(x => x.Token == jwt);
 }
