--- backend/DeltaShareApi/DeltaShareApi.http
@@ -1,6 +1,0 @@
-@DeltaShareApi_HostAddress = http://localhost:5041
-
-GET {{DeltaShareApi_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
