--- backend/DeltaShareApi/UserRole.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace DeltaShareApi
-{
-    public class UserRole
-    {
-        public Guid Id { get; set; }
-        [EmailAddress]
-        public string Email { get; set; } = default!;
-        public string Role { get; set; } = default!;
-    }
-}
