--- backend/DeltaShareApi/appsettings.json
@@ -1,12 +1,0 @@
-{
-    "ConnectionStrings": {
-        "DefaultConnection": "Host=localhost;Port=5432;Database=DeltaShare;Username=postgres;Password=postgres"
-    },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*"
-}
